requirements.txt
astroid==2.3.3
bleach==3.1.4
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9
docutils==0.16
harivansh-scripting-utilities==0.1.1
idna==2.9
isort==4.3.21
jeepney==0.4.3
keyring==21.2.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
pkginfo==1.5.0.1
pycparser==2.20
Pygments==2.6.1
pylint==2.4.4
readme-renderer==25.0
requests==2.23.0
requests-toolbelt==0.9.1
SecretStorage==3.1.2
six==1.14.0
tqdm==4.45.0
twine==3.1.1
urllib3==1.25.8
webencodings==0.5.1
wrapt==1.11.2
